+++ config/config.js
@@ -2,6 +2,7 @@
 //       You must provide your own.
 var token = require("./slack_token.js").token;
 var private_key = require("./service_account_key.js").key;
+var heltour_token = require("./heltour_token.js").token;
 
 try {
     var logging_token = require("./logging_token.js").token;
@@ -55,6 +56,14 @@ var config = {
                 "angborxley",
                 "prune2000",
             ],
+            "heltour": {
+                "token": heltour_token,
+                // TODO: this should be pulled from somewhere else so that
+                // config doesn't require updating each time a season changes
+                "base_endpoint": "https://www.lichess4545.com/api/", 
+                "season_tag": "4",
+                "league_tag": "team4545"
+            },
             "spreadsheet": {
                 "key": "1cTBOQNrbUw9-RWmrC9b6Yyi3i2MtIvUWuXNxzQmGfSs",
                 "serviceAccountAuth": {
@@ -118,6 +127,14 @@ var config = {
                 "lakinwecker",
                 "theino"
             ],
+            "heltour": {
+                "token": heltour_token,
+                // TODO: this should be pulled from somewhere else so that
+                // config doesn't require updating each time a season changes
+                "base_endpoint": "https://www.lichess4545.com/api/", 
+                "season_tag": "4",
+                "league_tag": "lonewolf"
+            },
             "spreadsheet": {
                 "key": "1WI4H9DWKSI-q6Omeqh28w1lKaqB0AAFV01wDiOeW6Ho",
                 "serviceAccountAuth": {
